clearing errors still working on gig errors
