+++ SlotMe.Services/SlotMe.Services.csproj
@@ -93,7 +93,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="GigServices1.cs" />
+    <Compile Include="GigService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SlotService.cs" />
     <Compile Include="TalentService.cs" />
@@ -112,11 +112,9 @@
       <Name>SlotMe.Models</Name>
     </ProjectReference>
   </ItemGroup>
-
   <ItemGroup>
     <WCFMetadata Include="Connected Services\" />
   </ItemGroup>
-
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
